Moved the validateSchemas method to its own class
