+++ prime-router/src/main/kotlin/azure/BlobAccess.kt
@@ -14,13 +14,6 @@ import com.azure.storage.blob.models.ListBlobsOptions
 import gov.cdc.prime.router.BlobStoreTransportType
 import gov.cdc.prime.router.Report
 import gov.cdc.prime.router.common.Environment
-import gov.cdc.prime.router.fhirengine.translation.hl7.FhirToHl7Converter
-import gov.cdc.prime.router.fhirengine.translation.hl7.FhirTransformer
-import gov.cdc.prime.router.fhirengine.translation.hl7.schema.ConfigSchemaReader
-import gov.cdc.prime.router.fhirengine.translation.hl7.schema.converter.ConverterSchema
-import gov.cdc.prime.router.fhirengine.translation.hl7.schema.fhirTransform.FhirTransformSchema
-import gov.cdc.prime.router.fhirengine.translation.hl7.utils.ConstantSubstitutor
-import gov.cdc.prime.router.fhirengine.utils.FhirTranscoder
 import org.apache.commons.io.FileUtils
 import org.apache.commons.io.FilenameUtils
 import org.apache.logging.log4j.kotlin.Logging
@@ -32,7 +25,6 @@ import java.net.URLDecoder
 import java.nio.charset.Charset
 import java.security.MessageDigest
 import java.time.Duration
-import javax.ws.rs.InternalServerErrorException
 
 const val defaultBlobContainerName = "reports"
 const val defaultBlobDownloadRetryCount = 5
@@ -119,7 +111,7 @@ class BlobAccess() : Logging {
     companion object : Logging {
         private const val defaultBlobDownloadRetryVar = "AzureBlobDownloadRetryCount"
         private val defaultEnvVar = Environment.get().blobEnvVar
-        private val defaultBlobMetadata by lazy {
+        val defaultBlobMetadata by lazy {
             BlobContainerMetadata.build(
                 defaultBlobContainerName,
                 defaultEnvVar
@@ -241,68 +233,6 @@ class BlobAccess() : Logging {
             val currentBlobItem: BlobItem,
             val previousBlobItemVersions: List<BlobItem>?,
         )
-
-        /**
-         * Downloads the content of each blob in the [directory] with [blobContainerInfo] takes the input file, runs
-         * the transform on it, and compares it to the output validating that they are the same.
-         */
-        fun validateSchemas(
-            directory: String,
-            blobContainerInfo: BlobContainerMetadata = defaultBlobMetadata,
-            schemaType: Report.Format,
-        ): Boolean {
-            val sourceContainer = getBlobContainer(blobContainerInfo)
-            val blobs =
-                listBlobs(directory, blobContainerInfo, false)
-            val transforms = blobs.filter { it.currentBlobItem.name.endsWith(".yml") }
-            transforms.forEach { currentTransform ->
-                val currentTransformName = currentTransform.currentBlobItem.name
-
-                val transformBlobClient = sourceContainer.getBlobClient(currentTransformName)
-                val transform = transformBlobClient.downloadContent()
-
-                val transformDirectoryPath = Regex("(.*/).*").find(currentTransformName)!!.groups[1]!!.value
-                val inputBlobClient = sourceContainer.getBlobClient(transformDirectoryPath + "input.fhir")
-                val input = inputBlobClient.downloadContent()
-                val inputBundle = FhirTranscoder.decode(input.toString())
-
-                if (schemaType == Report.Format.FHIR) {
-                    val configSchema = ConfigSchemaReader.readOneYamlSchema(
-                        transform.toStream(),
-                        FhirTransformSchema::class.java
-                    )
-                    val transformedInputBundle = if (configSchema is FhirTransformSchema) {
-                        FhirTransformer(configSchema).transform(inputBundle)
-                    } else {
-                        throw InternalServerErrorException("Transform schema $currentTransformName not a FhirTransform")
-                    }
-                    val output = sourceContainer.getBlobClient(transformDirectoryPath + "output.fhir").downloadContent()
-                    // Unfortunately need to do this to get it to be the same format.
-                    val expectedOutput = FhirTranscoder.decode(output.toString())
-                    // .equals is required here == does not compare it properly, despite what IntelliJ says
-                    if (transformedInputBundle.equals(expectedOutput)) {
-                        logger.error("Validation failed for transform $currentTransformName")
-                        return false
-                    }
-                } else {
-                    val converterSchema = ConfigSchemaReader.readOneYamlSchema(transform.toStream())
-                    val hl7Transform = FhirToHl7Converter(
-                        converterSchema as ConverterSchema,
-                        false,
-                        null,
-                        ConstantSubstitutor(),
-                        null
-                    ).convert(inputBundle)
-                    val output = sourceContainer.getBlobClient(transformDirectoryPath + "output.hl7").downloadContent()
-                    if (!hl7Transform.toString().trim().equals(output.toString().trim())) {
-                        logger.error("Validation failed for transform $currentTransformName")
-                        return false
-                    }
-                }
-            }
-
-            return true
-        }
 
         /**
          * Fetches all the blobs prefixed with [directory].  Azure stores blobs in a flat
@@ -449,7 +379,7 @@ class BlobAccess() : Logging {
          * If one exists for the container name and connection string, the existing one will be reused.
          * @return the blob container client
          */
-        private fun getBlobContainer(blobConnInfo: BlobContainerMetadata): BlobContainerClient {
+        fun getBlobContainer(blobConnInfo: BlobContainerMetadata): BlobContainerClient {
             return if (blobContainerClients.containsKey(blobConnInfo)) {
                 blobContainerClients[blobConnInfo]!!
             } else {
